Add public initializers to KeyPair structs
In this case, I find Swift's access levels a bit confusing. The
implicitly created initializer has the access level internal and can
therefore not be used outside of the library, even though all members
and the struct is marked as public. This is not a real issue as the
Sodium API takes PublicKey or SecretKey instances rather than KeyPairs.
But it would be nice to be able to re-use the KeyPair after persisting
keys without the need to extend it manually.