Disable Swift legacy version due to change to Swift 3
Although the syntax is Swift 3 already the Xcode project was left out
while switching to the new Swift version. This lead to Xcode prompting
the user to upgrade the syntax to Swift 3 upon opening the project. It
also lead to the project not building due to a dependency analysis
error:

  “Use Legacy Swift Language Version” (SWIFT_VERSION) is required to be
  configured correctly for targets which use Swift. Use the [Edit >
  Convert > To Current Swift Syntax…] menu to choose a Swift version or
  use the Build Settings editor to configure the build setting directly.

This patch fixes this behavior.