Use sodium_malloc(crypto_generichash_statebytes()) to alloc hash state struct
Swift no longer bridges C types with alignment sizes > 16 bytes, so instead
of trying to alloc the crypto_generichash_state struct in Swift, malloc it
using the recommended C functions and treat it as an UnsafeMutableRawPointer.

Wrapping it as a OpaquePointer is necessary for interacting with the other
C APIs, but we need to also keep the original UnsafeMutableRawPointer in order
to free the memory in deinit.